requirements.txt
sanic
tortoise-orm
passlib[argon2]
